\name{example_compounds}
\alias{example_compounds}
\docType{data}
\title{
	Example Compounds
}
\description{
	122 compounds in SDF format, stored as a list. Each element of the list is one line
	of text. This is just used in some unit tests.
}
%\usage{data(exmaple_compounds)}
\format{
  The format is:
 chr [1:12222] "3540" " OpenBabel06051210572D" "" ...
}
\keyword{datasets}
